refactor(psbt): extract finalize helpers into internal module
